Second stable version! Video call okayish, chat feature broken
