+++ refactor_video_call/global_state/index.js
@@ -1,0 +1,12 @@
+import { atom, selector } from "recoil";
+
+
+export const message = atom({
+  key: "message",
+  default: [],
+});
+
+export const menuOpened = atom({
+  key: "menuOpened",
+  default: 2,
+});
