+++ refactor_video_call/package.json
@@ -22,6 +22,7 @@
     "react-bootstrap": "^1.4.3",
     "react-dom": "17.0.1",
     "reactstrap": "^8.9.0",
+    "recoil": "^0.1.2",
     "socket.io-client": "^3.1.1"
   }
 }
