+++ refactor_video_call/pages/meet/[id].js
@@ -13,6 +13,8 @@ import "bootstrap/dist/css/bootstrap.css";
 import {isChrome, copyUrl} from "../../utils";
 import NavBar from "../../components/NavBar";
 
+import {black, silence} from "../../utils"
+
 const server_url =
   process.env.NODE_ENV === "production"
     ? ""
@@ -30,24 +32,6 @@ var socketId = null;
 var elms = 0;
 
 
-const silence = () => {
-    let ctx = new AudioContext();
-    let oscillator = ctx.createOscillator();
-    let dst = oscillator.connect(ctx.createMediaStreamDestination());
-    oscillator.start();
-    ctx.resume();
-    return Object.assign(dst.stream.getAudioTracks()[0], { enabled: false });
-  };
-
- const black = ({ width = 640, height = 480 } = {}) => {
-    let canvas = Object.assign(document.createElement("canvas"), {
-      width,
-      height,
-    });
-    canvas.getContext("2d").fillRect(0, 0, width, height);
-    let stream = canvas.captureStream();
-    return Object.assign(stream.getVideoTracks()[0], { enabled: false });
-  };
 
 
 const changeCssVideos = (main) => {
