+++ refactor_video_call/utils/browser_utils.js
@@ -1,0 +1,42 @@
+export const isChrome = function () {
+    if (typeof navigator ==="undefined" ) {
+      return;
+    }
+    let userAgent = (navigator && (navigator.userAgent || "")).toLowerCase();
+    let vendor = (navigator && (navigator.vendor || "")).toLowerCase();
+    let matchChrome = /google inc/.test(vendor)
+      ? userAgent.match(/(?:chrome|crios)\/(\d+)/)
+      : null;
+    // let matchFirefox = userAgent.match(/(?:firefox|fxios)\/(\d+)/)
+    // return matchChrome !== null || matchFirefox !== null
+    return matchChrome !== null;
+  };
+
+
+export const copyUrl = () => {
+    let text = window.location.href;
+    if (!navigator.clipboard) {
+      let textArea = document.createElement("textarea");
+      textArea.value = text;
+      document.body.appendChild(textArea);
+      textArea.focus();
+      textArea.select();
+      try {
+        document.execCommand("copy");
+        message.success("Link copied to clipboard!");
+      } catch (err) {
+        message.error("Failed to copy");
+      }
+      document.body.removeChild(textArea);
+      return;
+    }
+    navigator.clipboard.writeText(text).then(
+      function () {
+        message.success("Link copied to clipboard!");
+      },
+      () => {
+        message.error("Failed to copy");
+      }
+    );
+  };
+
